Improve FR Dashboard billing page

+++ db/migrate/20140520132150_add_clicks_counter_to_pledge.rb
@@ -1,5 +1,5 @@
 class AddClicksCounterToPledge < ActiveRecord::Migration
   def change
-    add_column :pledges, :clicks_count, :integer, default: 0
+    add_column :pledges, :clicks_count, :bigint, default: 0
   end
 end
